Template the SlurpIn packer to support Track and Hit inputs.
Use data type for the base pointer and perform all pointer arithmetic
with the data type, pass sizeof(data type) as scale argument to
SlurpIn.

Implement variations of SlurpIn that take T* as base argument. It also
takes scale as argument (though it could be hardcoded to
sizeof(T). The char* variant can later be removed.